+++ core/protocol/ide.ts
@@ -28,6 +28,8 @@ export type ToIdeFromWebviewOrCoreProtocol = {
   removeFile: [{ path: string }, void];
   showVirtualFile: [{ name: string; content: string }, void];
   openFile: [{ path: string }, void];
+  revealInExplorer: [{ path: string }, void];
+  revealInOS: [{ path: string }, void];
   openUrl: [string, void];
   runCommand: [{ command: string; options?: TerminalOptions }, void];
   getSearchResults: [{ query: string; maxResults?: number }, string];